dom/tests/mochitest/localstorage/test_cookieBlock.html

<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>cookie blocking test</title>

<script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

<script type="text/javascript">

function startTest()
{
  try {
    localStorage.setItem("blocked", "blockedvalue");
    ok(false, "Exception for localStorage.setItem, ACCESS_DENY");
  }
  catch (ex) {
    ok(true, "Exception for localStorage.setItem, ACCESS_DENY");
  }

  try {
    localStorage.getItem("blocked");
    ok(false, "Exception for localStorage.getItem, ACCESS_DENY");
  }
  catch (ex) {
    ok(true, "Exception for localStorage.getItem, ACCESS_DENY");
  }

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

SpecialPowers.pushPermissions([{'type': 'cookie', 'allow': false, 'context': document}], startTest);

</script>

</head>

<body>

</body>
</html>
